Implement Issue #4: Extract report status checking logic into useReportStatus hook
